@@ -11,14 +11,14 @@
     "combine"
   ],
   "engines": {
-    "node": ">=8.9.0"
+    "node": ">=8.12.0"
   },
   "dependencies": {
-    "hoek": "5.x.x"
+    "hoek": "6.x.x"
   },
   "devDependencies": {
     "code": "5.x.x",
-    "lab": "15.x.x"
+    "lab": "17.x.x"
   },
   "scripts": {
     "test": "lab -a code -t 100 -L",
